NetcodeGameObject.Step4
Created an object to spawn for each connected player <Player Prefab>

A simple Capsule is being used to visually represent the player character, and the prefab also has a Network Object Component attached.

Network Object Component enables the NetworkManager to spawn the player object when connecting as a client or host.